Discord_Bot/requirements.txt
discord-ext-voice-recv==0.4.2a145
discord.py==2.4.0
PyNaCl==1.5.0
requests==2.32.3
websockets==15.0.1